requirements.txt
streamlit>=1.51.0
pandas>=1.5.3
numpy>=1.24.2
finnhub-python>=2.4.18
